game.database.js szétszedése
